fix: helper 클래스 책임과 역할 분리
